+++ src/App.js
@@ -1,0 +1,10 @@
+import { Component } from "react";
+import Game from "./pages/Game";
+
+export class App extends Component {
+    render() {
+        return (<div>
+            <Game />
+        </div>);
+    }
+}
