+++ src/components/Square/index.js
@@ -1,0 +1,3 @@
+import Square from './Square';
+
+export default Square;
