+++ src/components/Switch/index.js
@@ -1,0 +1,3 @@
+import Switch from './Switch';
+
+export default Switch;
