+++ src/pages/Game/index.js
@@ -1,0 +1,3 @@
+import Game from "./container";
+
+export default Game;
